Add files via upload
This is my latest and most complete set of images for replacing the gbz35 theme backgrounds on any raspberry pi with a 640x480 screen. I made this set solely for use on my Experimental Pi piboy dmg. they will work on other screen sizes and aspects, but they will likely stretch and look less than perfect. the file structure is identical to that of the installed system files to make for easy transfer. If for some reason you have changed the file name of one of your emulator folders, for example, 'GenH' to 'GenesisH' you can still use the files from here, but you would have to rename the folder to suit your device (I try to not change from defaults for this very reason) or just copy the single folder contents and not the full folder. to make life simple, you will find some folders on here replicate each other in an attempt to cater to some of those people who have different named folders created by more obscure emulators. If you find a common machine thats not showing a custom background that you know is in this set, its a 99 percent bet that its your machine have an oddly named folder causing it. just copy the relevant files from this set into the one causing you trouble. there is an image file here showing the primary backgrounds and buttons that are included. it does Not show alternative images for all the systems that went under different names for different countries (genesis/megadrive for example). They are on here in the relevant folders though.