Add files via upload
I spotted a small error with the sg1000 background image on my machine earlier. It was from the first set (before I moved the dates up a bit). The release year on the old v1 set was too low and impinged on the quick menu line. I don't remember seeing it like this until now and there is every possibility I uploaded the correct file in the main zip, but just in case I didn't, here is the correct image with the date in the right place.